5786: make label a computed property and removing extra observer in select field
